+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 1. Improved the Order model to include the Model Extension concern
 2. Improved the Order Item model to include the Model Extension concern
 3. Improved the Order Adjustment model to include the Model Extension concern
+
+## v0.3
+1. Improved the Order Item model to add the ::order! method
+2. Improved the Order Adjustment model to rename the #adjusted_by scope to #adjusted_is
+3. Improved the Order Adjustment model to add the ::adjust! method
+4. Improved the Ruby Gem Specification to depend on [unidom-common](https://github.com/topbitdu/unidom-common) v0.9